Escape '%' in unit description
'%' has special meaning in systemd unit files. It has to be properly escaped to avoid errors.
Doing so by replacing '%' with '%%' in the description field.

The description field is populated with the cron command line by systemd-crontab-generator.py
and can therefore contain '%', like e.g. in Debian's mdadm cron command.