Prune old cron-*.timer units in cron-update
We don't touch .service, since this may be interesting to the
administrator if it failed, but .timers enter a "failed" state when
they're removed, and need a manual prune, and they naturally can't
"fail" otherwise.

This is especially an issue when users are editing their crontabs,
and leave behind useless drivel that only the administrator can clean up.