Replace openssl (-lcrypto) with libmd (-lmd)
We only use it for MD5 calculation, so the huge libssl dep is kinda
wasted. libmd is already used by dpkg &a. so it's not an additional
dependency. Also, going from
  $ ldd out/build/bin/systemd-crontab-generator
          linux-vdso.so.1 (0x00007fff1a9f0000)
          libcrypto.so.3 => /lib/x86_64-linux-gnu/libcrypto.so.3 (0x00007ff13e4d1000)
          libstdc++.so.6 => /lib/x86_64-linux-gnu/libstdc++.so.6 (0x00007ff13e2b7000)
          libm.so.6 => /lib/x86_64-linux-gnu/libm.so.6 (0x00007ff13e1d8000)
          libgcc_s.so.1 => /lib/x86_64-linux-gnu/libgcc_s.so.1 (0x00007ff13e1b8000)
          libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007ff13dfd7000)
          /lib64/ld-linux-x86-64.so.2 (0x00007ff13e987000)
to
  $ ldd out/build/bin/systemd-crontab-generator
          linux-vdso.so.1 (0x00007ffcf714f000)
          libmd.so.0 => /lib/x86_64-linux-gnu/libmd.so.0 (0x00007f4e44226000)
          libstdc++.so.6 => /lib/x86_64-linux-gnu/libstdc++.so.6 (0x00007f4e4400c000)
          libm.so.6 => /lib/x86_64-linux-gnu/libm.so.6 (0x00007f4e43f2d000)
          libgcc_s.so.1 => /lib/x86_64-linux-gnu/libgcc_s.so.1 (0x00007f4e43f0d000)
          libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f4e43d2c000)
          /lib64/ld-linux-x86-64.so.2 (0x00007f4e44267000)
means we also go from
  -rw-r--r-- 1 root root 4.5M 10-23 19:52 /lib/x86_64-linux-gnu/libcrypto.so.3
to
  -rw-r--r-- 1 root root  47K 2022-07-10  /lib/x86_64-linux-gnu/libmd.so.0
which means our RSS and pagefaults go down by half
(according to /bin/time), and we save 1-2ms/run (according to bash time).

The libmd interface is also simpler and infallible, unlike the
openssl interface (and we were using the "deprecated" EVP_md5() API),
and we can use its own binary-string formatter instead of rolling our own.

I had wanted to use libmd originally but couldn't drive it correctly for
some reason (probably because of the confusingly-named functions).